Delete a pizza order

+++ routes/web.php
@@ -26,8 +26,17 @@ Route::get('/', function () {
 
 use App\Http\Controllers\PizzaController;
 
-Route::get('/pizzas', [PizzaController::class,'index']);
-Route::get('/pizzas/create', [PizzaController::class,'create']); //Always put this above the {id} route to avoit confusing in url
-Route::post('/pizzas', [PizzaController::class, 'store']); //New record into the DB
-Route::get('/pizzas/{id}', [PizzaController::class,'show']);
-Route::delete('/pizzas/{id}', [PizzaController::class,'destroy']);
+//->name('pizza.index') become a variable for the router path
+
+Route::get('/pizzas', [PizzaController::class, 'index'])->name('pizzas.index')->middleware('auth'); //->middleware set only auth user can access the page
+Route::get('/pizzas/create', [PizzaController::class, 'create'])->name('pizzas.create'); //Always put this above the {id} route to avoit confusing in url
+Route::post('/pizzas', [PizzaController::class, 'store'])->name('pizzas.store'); //New record into the DB
+Route::get('/pizzas/{id}', [PizzaController::class, 'show'])->name('pizzas.show')->middleware('auth');
+Route::delete('/pizzas/{id}', [PizzaController::class, 'destroy'])->name('pizzas.destroy')->middleware('auth');
+
+Auth::routes([
+    'register' => false //This turn off the register route in the whole project, no longer be accessible
+    //Switch it to true, if you wanna register new admin
+]);
+
+Route::get('/home', [App\Http\Controllers\HomeController::class, 'index'])->name('home');
